requirements_CombiCR.txt
# Automatically generated by https://github.com/damnever/pigar.

biopython == 1.79
matplotlib == 3.3.1
numpy == 1.17.0
pandas == 1.1.5
scipy == 1.5.2
seaborn == 0.11.2
tqdm == 4.64.1
